Backend/requirements.txt
fastapi
uvicorn

# database libs
sqlalchemy
asyncpg
psycopg2

# env libs
python-dotenv

# email validation libs
pydantic[email]

# hashing password lib
passlib[bcrypt]

# fastapi-mail
fastapi-mail

# JWT tokens
python-jose